thermal: qcom-spmi-temp-alarm: Add support for Deep Sleep
Add suspend and resume functionality for qcom-spmi-temp-alarm to support
DeepSleep. For a DeepSleep entry/exit, perform the same steps that are
followed for Hibernation entry/exit.

From kernel 6.1 onward we are not using power_state.c due to
GKI compliance. So, we are updating the state via
pm_suspend_target_state.

Change-Id: Icf671d8d1144761b571a404419960bfbdae197ca
Signed-off-by: Brindha T <quic_brint@quicinc.com>