i3c-master-msm-geni: Fix IBI controller stuck issue due to HW BUG
If an error such as a NACK or timeout is detected from
the slave side, the I3C driver initiates a STOP sequence
on the bus by sending cancel or abort commands. When the
GENI controller is IDLE, the IBI controller executes a
STOP-only opcode on the bus. During this process, if the
slave issues an IBI the IBI event is received. However,the
IBI controller's finite state machine (FSM) cannot handle
STOP command execution when the bus is not IDLE, which
causes the controller to hang.

Due to HW BUG, SWA was introduced by setting the
stretch bit to 1. This keeps SCL LOW, preventing the slave
from generating an IBI during STOP execution. This SWA is
fixed in the HW from core version >=4. So, added required
changes as per qup core version.

Change-Id: Iee35a5b67d306fffd43d384d4edb3c767eb757fc
Signed-off-by: Anil Veshala Veshala <quic_aveshala@quicinc.com>
Signed-off-by: Krishna Chaithanya Reddy G <quic_kgangapu@quicinc.com>