gunyah: Add support to handle VM err-fatal
Add support to handle secondary VM's error-fatal scenarios.
The patch starts listening to the VM_EXITED notifications to
detect if the secondary VM has come down.

Implement sysfs interface recovery.This gives the user-space
to control the behaviour upon SVM error-fatal. It can choose
to panic the entire system or can opt to get notified, such
that it can try to restart later.

Change-Id: I57540cd4375afb8420e45c14770b287d9fb9bdba
Signed-off-by: Peng Yang <quic_penyan@quicinc.com>