gunyah: Create a vcpu workqueue only if the rebootable VM is still running
When we start supporting VM restart, we should be able to clean up the
VM resources when the VM shuts down. With the keep-running flag, we don't
do any cleanup if the VM shts down failing the next boot of the VM. The
reason we need the keep-running flag is for the scenario when qcrosvm
aborts or if the PVM panics where we want to forcibly run the vcpus to
completion. We still want to keep this feature alongside the VM shutdown
case. Thus first check if the VM is actually running before we forcibly
run the vcpus.

Change-Id: I11be990b5c66a24b389325e1c88e89db84495ac0
Signed-off-by: Prakruthi Deepak Heragu <quic_pheragu@quicinc.com>
Signed-off-by: Cong Zhang <quic_congzhan@quicinc.com>