spi: spi-msm-geni: Fix clock instability by setting index before rate
SSC QUP may experience clock instability if the clock rate is changed
before the parent clock has fully switched. This can result in
incomplete transitions of the clock parent and, in some cases, may
cause the system to hang.

To address this, configure the clock selection index and divider before
setting the clock rate.

Change-Id: Iacad1dff79f0050a34cf41788f2fab48a53d2ae8
Signed-off-by: Zahir Shabbir Khan <quic_zshabbir@quicinc.com>
Signed-off-by: Maramaina Naresh <quic_mnaresh@quicinc.com>