serial: msm_geni_serial: Add one wire UART feature
One wire UART minimizes the number of GPIO pins needed
for data transfer by using a single data line for
communication between the host controller and the
device. This interface operates in half-duplex, meaning
that either the host or the device can drive the bus
at any given time but not both simultaneously.

To enable the one-wire UART feature, ensure firmware is
supported. Add the compatible string qcom,msm-geni-one-wire-serial
and introduce the necessary device tree settings.

Change-Id: If0fdb9e2afc72302bb9f6459124541e646a3c88a
Signed-off-by: Krishna Chaithanya Reddy G <quic_kgangapu@quicinc.com>