phy: eusb: phy-qcom-snps-eusb2: Get proper usb psy for second phy
The reference to power supply for setting ICL is obtained by
referencing "usb" name. But when ICL changes need to be done on
second port, the reference of "usb-2" has to be taken. Else the ICL
setting would be done only on first port irrespective of which port
is active. Use "qcom,port-num" property to identify if the phy under
usage is secondary and get appropriate power supply reference.

Change-Id: I509a1190eb58216c253d451523425cc5667167c7
Signed-off-by: Krishna Kurapati <krishna.kurapati@oss.qualcomm.com>