serial: qcom-geni: Add support for Suspend-to-RAM (S2R)
Add support for the Suspend-to-RAM (S2R) feature,where the SoC is powered
down and its state is preserved in DRAM. Upon resume, the hardware context
is lost and must be re-initialized, while the software context remains
intact.

The serial engine is reconfigured during resume, similar to the initial
probe sequence. The PM_SUSPEND_MEM flag is used to detect S2R transitions.

Change-Id: I37710a55b91daea852a60a8ba06fa0bb87839c0d
Signed-off-by: Praveen Talari <quic_ptalari@quicinc.com>