usb: dwc3: dwc3-msm-core: Remove redundant driver string checks
Commit 7fe440ccffaa1 ("usb: dwc3: msm: Allow prepare/complete
callbacks only for msm-dwc3 child") The fix introduced a check for
the parent driver name in host/device mode. However, in host mode,
the dev argument passed to host_prepare and host_complete is xhci,
with the parent being dwc and not dwc3-msm as initially expected.
This led to a failure in runtime suspend for host mode.

Fix this by removing redundant checks in the `dwc3_host_complete`
and `dwc3_host_prepare` functions that verify the device parent string.

Change-Id: I189d09b5a0ef88028cc455ab78c5c4fbb1daee70
Signed-off-by: Rajashekar kuruva <quic_kuruva@quicinc.com>